requirements.txt
azure-ai-inference>=1.0.0
azure-core>=1.29.5
python-dotenv>=1.0.0